Change Frequency Table data structure
Usage of array of array was enough in order to work with two kinds of
data so far (xi and fi).

To make it easier to add other kinds of data to it, it is better to have
an object, so it's possible to use an unique identifier for each kind.